backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
apscheduler
anthropic
moorcheh-sdk
twilio
python-dotenv
pydantic
pydantic-settings
httpx
